+++ app/app.js
@@ -16,13 +16,13 @@ const root = require('./src/apis/root');
 const review = require('./src/apis/review');
 const home = require('./src/apis/circle-home');
 const schedule = require('./src/apis/schedule');
-const clubList = require('./src/apis/club-list');
+const clubList = require('./src/apis/club');
 
 app.use('/api/board', board);
 app.use('/api', root);
 app.use('/api/club/review', review);
 app.use('/api/club/home', home);
 app.use('/api/club/schedule', schedule);
-app.use('/api/clubList', clubList);
+app.use('/api/club/list', clubList);
 
 module.exports = app;
